Fix register DAY
Python tm_day is 0-6 0=Monday.
DS1307 DAY register is 1-7.

From datasheet:
The day-of-week register increments at midnight.
Values that correspond to the day of week are user-defined but must be
sequential (i.e., if 1 equals Sunday, then 2 equals Monday, and so on.)

Also don't set time when ram is written.